fixed lots of server intigration

+++ server/models/favorites.model.js
@@ -2,10 +2,11 @@ const query = require("../config/mysql.config");
 
 async function addFavorite(res, userId, recipeId) {
   try {
-    await query("INSERT INTO favorites (user_id, recipe_id) VALUES (?, ?)", [
-      userId,
-      recipeId,
-    ]);
+    const [check] = await query("SELECT * FROM favorites WHERE favorites.user_id = ? AND favorites.recipe_id = ?", [userId, recipeId]);
+    if (check) {
+      return res.send({ success: false, data: null, error: "already in your favorites" });
+    }
+    await query("INSERT INTO favorites (user_id, recipe_id) VALUES (?, ?)", [userId, recipeId]);
     return res.send({ success: true, data: null, error: null });
   } catch (error) {
     console.log(error);
@@ -19,10 +20,7 @@ async function addFavorite(res, userId, recipeId) {
 
 async function removeFavorite(res, userId, recipeId) {
   try {
-    await query(`DELETE FROM favorites WHERE user_id = ? AND recipe_id = ?`, [
-      userId,
-      recipeId,
-    ]);
+    await query(`DELETE FROM favorites WHERE user_id = ? AND recipe_id = ?`, [userId, recipeId]);
     return res.send({ success: true, data: null, error: null });
   } catch (err) {
     return res.send({
